requirements.txt
docx==0.2.4
docx2pdf==0.1.8
docxcompose==1.4.0
docxtpl==0.18.0
moviepy==1.0.3
openpyxl==3.1.5
packaging==24.1
pandas==2.2.2
pillow==10.2.0
PyPDF2==3.0.1
python-dateutil==2.9.0.post0
python-decouple==3.8
python-docx==1.1.2
python-pptx==1.0.2
python-telegram-bot==21.4
openai==1.41.0
matplotlib==3.9.2
gTTS==2.5.3
APScheduler==3.10.4
arabic-reshaper==3.0.0
asgiref==3.7.2